QnA & Saved Jobs
Made it so that the quick apply button on saved jobs disappears if you have already applied.
Added QnA on the job view page. This allows the user to post questions on the job listing. Added features so far:
Question submitting
Question editing
Vote system so that questions with most votes(most helpful questions) will be at the top of the QnA.