QnA & Saved Jobs: Both Completed
Finished my tasks.
Since it was made clear that only one person is allowed to commit and merge to main I will put the remainder of my work in this branch.
While I was just finishing the page I was working on, someone decided to split it in two on the main branch without informing me of this. Due to this it is now the responsibility of the person who merges this branch into main to split what I've done to fit the two new pages that replace this one. This is because someone didn't think I needed to be informed of this change until I had completed everything; so it is now the responsibility of the one who made this change and is merging to figure out how to solve this problem they created for themselves :)